+++ packages/core/src/runbook/index.ts
@@ -17,6 +17,8 @@ export { SessionService } from './session-service.js';
 export { ExecutionLifecycleService } from './execution-lifecycle-service.js';
 export { compileRunbookToMachine, runbookSetup, MAX_FILE_ITERATIONS } from './compiler.js';
 export type { RunbookMachine } from './compiler.js';
+export { runRetryHook } from './retry-hook.js';
+export type { RetryHookResult, RetryHookSuccess, RetryHookError } from './retry-hook.js';
 export {
   executeCommand,
   executeCommandWithPolicy,
